+ resolved #15 + resolved #16 + resolved #19 + resolved #20
