+++ pom.xml
@@ -7,13 +7,13 @@
     <groupId>tech.ynfy</groupId>
     <artifactId>ynfy-tool-httpconnect</artifactId>
     <name>${project.groupId}:${project.artifactId}</name>
-    <version>1.0.5</version>
+    <version>1.1.0-SNAPSHOT</version>
     <description>A Http Util for Java</description>
     <url>https://github.com/ynfy-tech/ynfy-tool-httpconnect</url>
 
     <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
     </properties>
 
     <!-- open source info -->
